One-Arm Chinup training tracker
Since it's hard for me to count to 20 or even 10.

Command line program for my one-arm chinup training's tracking (20 reps within 20 min).

when compiled and executed:
"training.txt" is made/opened
start time is displayed on terminal and logged to the file
stopwatch for first rep starts (along with a colored "progress bar" to show at a glance the time spent on the current rep)
when you tap any key: the stopwatch stops, and the time and rep# are logged to the file

this repeats until the 20th rep, after which the finish time is logged to the file and the program ends.

Makes use of threads/synchronisation/concurrency, which i learned in the semester just before making this, for my convenience as qwell as to practise for the semester final exam.

The logging to the file is done with formatting and coloring using ANSI escape codes, so that writing "cat training.txt" would show a clear, clean, colored output.